strong refactor as bionc_numpy and bionc_casadi
