+++ bionc/model_creation/natural_segment_template.py
@@ -4,7 +4,7 @@ from .natural_axis_template import AxisTemplate
 from ..model_computations.biomechanical_model import BiomechanicalModel
 
 # from ..utils.natural_coordinates import SegmentNaturalCoordinates
-from bionc.math_interface.protocols.natural_coordinates import SegmentNaturalCoordinates
+from ..protocols.natural_coordinates import SegmentNaturalCoordinates
 from .marker_template import MarkerTemplate
 from .protocols import Data
 from ..model_computations.natural_segment import NaturalSegment
@@ -60,7 +60,7 @@ class NaturalSegmentTemplate:
         SegmentNaturalCoordinates
         The Segment Natural Coordinates Q (12 x n_frames)
         """
-        from ..math_interface.using_numpy import SegmentNaturalCoordinates
+        from ..bionc_numpy import SegmentNaturalCoordinates
 
         self.Q = SegmentNaturalCoordinates.from_components(
             u=self.u_axis.to_axis(data, kinematic_chain).axis()[:3, :],
